requirements.txt
torch>=2.0.0
transformers>=4.31.0
sentence-transformers
Pillow
numpy
scikit-learn
pyyaml
matplotlib
tqdm
bitsandbytes
accelerate
